refactor: extract GitHub operations into separate module, expand server route tests
- Move :TgPr and :TgIssue from init.lua (857 lines → 480 lines) to
  lua/telegram/github.lua
- Expand server.test.ts from 1 test to 17, covering all major routes
  (health, chats, messages, send, edit, delete, forward, open/close chat)
- Includes date-based filtering test for /messages endpoint

Closes #71, #73
